Moves serialization tests into a separate test file
