+++ lib/Macaroons/Serializers/JSONSerializer.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Macaroons\Serializers;
+
+class JSONSerializer extends BaseSerializer
+{
+}
