+++ tests/PacketTest.php
@@ -3,7 +3,7 @@ namespace Macaroons\Tests;
 
 use Macaroons\Utils;
 use Macaroons\Macaroon;
-use Macaroons\Packet;
+use Macaroons\Serializers\Packet;
 
 class PacketTest extends \PHPUnit_Framework_TestCase
 {
